test: adopt new testScope for multus cni tests
